# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Global Tongue API Documentation",
        "description": "Heirarchy of API | Courses -> Levels -> Units -> Lessons -> Exercises",
        "contact": {
            "email": "gemeelijah@gmail.com"
        },
        "license": {
            "name": "www.globaltongueedu.com",
            "url": "https://globaltongueedu.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/public/api/lessons": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get Lesson List",
                "description": "Get Lesson List as Array",
                "operationId": "LessonIndex",
                "responses": {
                    "200": {
                        "description": "Get Lesson List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/lessons/{id}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Show Lesson Details",
                "description": "Show Lesson Details",
                "operationId": "showLesson",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Lesson Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/lessons/by/{unit_id}": {
            "get": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Get Lessons by Unit ID",
                "description": "Get list of lessons associated with a specific unit",
                "operationId": "getLessonsByUnitID",
                "parameters": [
                    {
                        "name": "unit_id",
                        "in": "path",
                        "description": "ID of the unit",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lessons for the specified Unit ID"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "No Lessons found for the specified Unit ID"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/lessons/create": {
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create New Lesson",
                "description": "Create New Lesson",
                "operationId": "storeLesson",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "level_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "unit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "lesson_title": {
                                        "type": "string",
                                        "example": "Essential Pronunciation: -ch, -h, -ll, -ñ"
                                    },
                                    "lesson_type": {
                                        "type": "string",
                                        "example": "voice"
                                    },
                                    "lesson_cover": {
                                        "type": "string",
                                        "example": "lesson_image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create New Lesson"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}